+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - successfully parsed file to verbose output
 
 ### Changed
+- improve syntax error message
 - continue running all the way through on syntax error
 
 [0.49.2](https://github.com/jshwi/docsig/releases/tag/v0.49.2) - 2024-04-21